requirements.txt
openai
python-dotenv
pandas
matplotlib
seaborn